bug: use correct embedding field in vector search
